feat: Add react-router-dom to project dependencies
This commit adds the `react-router-dom` package to the project's dependencies in both the `package.json` and `package-lock.json` files. The version `6.23.0` of `react-router-dom` is specified.

The purpose of this change is to enable the usage of React Router in the project, allowing for client-side routing and navigation. This will be beneficial for implementing multiple pages and routes in the application.